refactor creation code into WIP system2
